+++ src/app/components/Dashboard/ArticleDashboard.js
@@ -1,6 +1,8 @@
 import React from 'react';
 
-const ArticleDashboard = () => <div>ArticleDashboard</div>;
+const ArticleDashboard = () => (
+  <div>ArticleDashboard</div>
+);
 
 // eslint-disable-next-line import/no-unused-modules
 export default ArticleDashboard;
